image_pipe 코드 분리 및 버그 해결
